Logout redirects and signs user out on each page
